requirements.txt
face-recognition
opencv-python
amcrest
scipy
openai-whisper
elevenlabs
sounddevice
python-dotenv
numpy
